+++ src/components/ThemeProvider/index.tsx
@@ -1,3 +1,4 @@
-// ThemeProvider from copy-ui @ 2025-10-05.
+// ThemeProvider from copy-ui @ 2025-10-08.
 
+export type { ColorName, ColorNo } from "./ThemeProvider";
 export { ThemeContext, ThemeProvider, useTheme } from "./ThemeProvider";